Adding component modes and Button component
Components can now be in one of three modes: active, inactive, or hovering. The gui creator can specify different component states for each mode. Adding Button component, which uses the new multi-mode functionality. The button is active when it is being pressed, hovering when the mouse is hovering over it and it is not being pressed, and inactive otherwise. The Button component itself defines the state transitions. The gui creator defines what the button looks like in each state. The gui creator also defines a listener that gets notified when the button is clicked.